# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface EmailSummary {
  id: string;
  from: string;
  subject: string;
  date: string;
  read: boolean;
  preview: string;
}

export interface EmailList {
  emails: EmailSummary[];
  total: number;
  page: number;
  totalPages: number;
}

export type EmailDetailAttachmentsItem = {
  filename: string;
  size: number;
  contentType: string;
};

export interface EmailDetail {
  id: string;
  from: string;
  to: string;
  subject: string;
  date: string;
  htmlBody?: string;
  textBody?: string;
  read: boolean;
  attachments?: EmailDetailAttachmentsItem[];
}

export interface Domain {
  id: number;
  name: string;
  createdAt: string;
  active: boolean;
}

export interface DomainList {
  domains: Domain[];
}

export interface AddDomainBody {
  name: string;
}

export interface ErrorResponse {
  error: string;
}

export interface EmailStats {
  totalEmails: number;
  unreadEmails: number;
  lastChecked?: string;
}

export type ListEmailsParams = {
  /**
   * The email address to fetch emails for
   */
  address: string;
  page?: number;
};

export type GetEmailParams = {
  address: string;
};

export type GetEmailStatsParams = {
  address: string;
};